requirements.txt
datasets
sentence-transformers
chromadb
openai
python-dotenv